youtube-analytics-dashboard/backend/requirements.txt
fastapi
uvicorn
pandas
google-generativeai
python-multipart
pydantic
